Add functionality to forward missed Activity Lifecycle callbacks
We introduce a decorator over Context and ApplicationContext which will account for any Activity lifecycle callbacks which are missed between The time the Application is created, and the time any kit may be initialized and created. It was found that some SDKs may relied upon callbacks occurring before the kit completed initialization, so this method was necessary to replay or forward these callbacks to the SDK upon initialization. This method also allows for us to guarantee that Activity lifecycle remain atomic

As a result of the changes to KitContext, we had to add a constructor testing flag, in order to make the class testable