Gracefully handle late SDK initialization
We document that the SDK must be started within Application#onCreate, but
it may come to pass that that is impossible to overlooked. This commit
is geared towards handling the case where we're initialized later in the lifecycle.

Apps that require starting the SDK later can now pass an Activity to start, which
we will use to kick off our session management.