Fix session-end message creation for orphan sessions
This addresses a bug that occurs when the app is frequently
force-closed by the user prior to upload. On app-open, the SDK
will query for old sessions, and create a session-end event for them.
The SDK will then upload the session-end message, and delete the session
from the database.

Prior to this fix, if the app continued to be force-closed prior to upload
and subsequent deletion of the session, the SDK would continually create
new session-end messages for the same, undeleted sessions. This fix introduces a
flag to the session table to track session-end creation.