Persist user identities in specific order
This commit addresses an issue with some downstream consumers of identities, kits in particular,
which need to be given the customer ID first. This is largely a mitigation for kits that have not
yet updated to the new Identity APIs to detect login/logout/modify such that they can control the
identity ordering themselves.