Defer logPushRegistration()'s modify request
Previously, MParticle.logPushRegistration() invocations would result in a failed modify() requests when an MParticleUser was not preset. This changes that behavior in that now we will defer that modify request until an MParticleUser is present. This behavior is most salient and repeatable when an MParticleOptions instance containing a pushRegistration field is used to start the SDK on a fresh install