Fix automatic push token fetch
A bug was introduced which caused any invocation of PushRegistrationHelper.requestInstanceId() to often defer fetching a new Token when it shouldn't have. This was due to an incorrect refactor of the mechanism which limits a new Token to be fetched only when a version upgrade or OS upgrade was detected. Although this bug affected calls to MParticle.Messaging().enablePushNotifications(), it did not affect calls toMParticle.logPushRegistration()