Replace manual instrumentation with Aspects
Support for SdkListener.onApiClassed() requires every API method to be manually instrumented with an InternalListenerManager.onApiCalled() call. This replaces that manual instrumentation with an AspectJ implementation which will insert similar hooks automatically, in the pre-compile stage.

Tests are also encluded to enforce which API methods we want to be instrumented with these hooks. The Aspect pointcut is currently set to look at class with the @ApiClass annotation and instrument it's public methods, minus the getters/is classes. There is one method Cart.clear(): Cart which does not fit the main rule, so that method is still instrumented manually