Make ApplicationContextWrapper replay lifecycle callbacks asynchronously
Since Kits are initialized on a delay, the ApplicationContextWrapper collects lifecycle callbacks that occur between the apps initialization and when an ApplicationStateListener is attached, at which point the collected events are replayed. This allows our partner SDKs who need to observer lifecycle callbacks from application initialization to function properly. We found that some SDKs register for lifecycle callbacks within their initialization, and it was problematic to invoke lifecycle callbacks before their initialization was complete. This change makes it so that lifecycle callbacks are no longer replayed synchronously when the listener is registered, rather the replay is queued in the Looper via a Handler.