Check MParticle is not null when Session ends
In certian conditions an NPE is thrown when MParticle is dereference before a Session ends. We run a Handler with a delayed callback which will check if the application has gone into the background, and end the current session as well as disable location tracking, if it had been enabled. This call the disable location tracking did not check if the main MParticle instance was non-null, so when the SDK was dereferenced before the Handler executed, the call would throw and uncaught exception