Refactor existing Kit end-to-end tests
Small improvement to the Kit integration testing. Previously, you would register a test kit with a String which was supposed to be the qualified class name. This requirement was ambiguous, so I replaced it with a specific abstract base test kit implementation. Base test kit also implements getInstance(), which was an unclear reaquirement, but the only way a kit would be returned via MParticle#getKitInstance()