fix: Hardcode macos version for Github Actions
The 3rd party Github Action we use to instantiate Android emulators broke with the most recent macos release. We were able to resolve the issue by hardcoding the macos version rather than relying on the most recent release. We also hardcoded the dependency version on the emulator action for additional stability in line with best practices.