feat: Add configMaxAge to MParticleOptions
Set a maximum age threshold for the locally cached config. When a store config surpasses the threshold, it will be purged on the next SDK start, and the SDK will operate without a config until one is returned successfully from the server. configMaxAge is set on startup, in seconds, by calling MParticleOptions.Builder#configMaxAge(Long)