fix: modify() call throws an exception when user is not present (#195)
When the AdId updates, we will generate a modify() request interanlly with the current user. When the current user is not present/null, this request currently fails with an error log or a thrown exception, if the SDK is running in Development mode. In order to avoid losing this modify request or causing an exception, we will defer the modify request until a user is present, if it is not already